Move VirtualInventory#ItemSignature to its own class
