+++ core/src/main/java/github/nighter/smartspawner/spawner/data/SpawnerFileHandler.java
@@ -2,6 +2,7 @@ package github.nighter.smartspawner.spawner.data;
 
 import github.nighter.smartspawner.SmartSpawner;
 import github.nighter.smartspawner.spawner.data.storage.SpawnerStorage;
+import github.nighter.smartspawner.spawner.properties.ItemSignature;
 import github.nighter.smartspawner.spawner.properties.SpawnerData;
 import github.nighter.smartspawner.spawner.properties.VirtualInventory;
 import github.nighter.smartspawner.Scheduler;
@@ -229,7 +230,7 @@ public class SpawnerFileHandler implements SpawnerStorage {
 
                 VirtualInventory virtualInv = spawner.getVirtualInventory();
                 if (virtualInv != null) {
-                    Map<VirtualInventory.ItemSignature, Long> items = virtualInv.getConsolidatedItems();
+                    Map<ItemSignature, Long> items = virtualInv.getConsolidatedItems();
                     List<String> serializedItems = ItemStackSerializer.serializeInventory(items);
                     spawnerData.set(path + ".inventory", serializedItems);
                 }
