+++ core/src/main/java/github/nighter/smartspawner/spawner/gui/sell/SpawnerSellConfirmUI.java
@@ -1,5 +1,6 @@
 package github.nighter.smartspawner.spawner.gui.sell;
 
+import github.nighter.smartspawner.spawner.properties.ItemSignature;
 import net.kyori.adventure.text.Component;
 import github.nighter.smartspawner.SmartSpawner;
 import github.nighter.smartspawner.language.LanguageManager;
@@ -10,7 +11,6 @@ import github.nighter.smartspawner.spawner.gui.layout.GuiLayout;
 import github.nighter.smartspawner.spawner.lootgen.loot.EntityLootConfig;
 import github.nighter.smartspawner.spawner.lootgen.loot.LootItem;
 import github.nighter.smartspawner.spawner.properties.SpawnerData;
-import github.nighter.smartspawner.spawner.properties.VirtualInventory;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.entity.EntityType;
@@ -164,7 +164,7 @@ public class SpawnerSellConfirmUI {
 
     private ItemStack createSpawnerInfoButton(Player player, SpawnerData spawner, Map<String, String> placeholders) {
         // Build loot item components for {loot_items} placeholder
-        Map<VirtualInventory.ItemSignature, Long> storedItems = spawner.getVirtualInventory().getConsolidatedItems();
+        Map<ItemSignature, Long> storedItems = spawner.getVirtualInventory().getConsolidatedItems();
         List<Component> lootComponents = buildSellInfoLootComponents(spawner, storedItems);
 
         // Prepare the meta modifier consumer
@@ -200,9 +200,9 @@ public class SpawnerSellConfirmUI {
         return spawnerItem;
     }
 
-    private List<Component> buildSellInfoLootComponents(SpawnerData spawner, Map<VirtualInventory.ItemSignature, Long> storedItems) {
+    private List<Component> buildSellInfoLootComponents(SpawnerData spawner, Map<ItemSignature, Long> storedItems) {
         Map<Material, Long> materialAmountMap = new HashMap<>();
-        for (Map.Entry<VirtualInventory.ItemSignature, Long> entry : storedItems.entrySet()) {
+        for (Map.Entry<ItemSignature, Long> entry : storedItems.entrySet()) {
             Material material = entry.getKey().getTemplateRef().getType();
             materialAmountMap.merge(material, entry.getValue(), Long::sum);
         }
@@ -227,9 +227,9 @@ public class SpawnerSellConfirmUI {
                         "button_sell_info.loot_items", material, formattedAmount, chance));
             }
         } else {
-            List<Map.Entry<VirtualInventory.ItemSignature, Long>> sortedItems = new ArrayList<>(storedItems.entrySet());
+            List<Map.Entry<ItemSignature, Long>> sortedItems = new ArrayList<>(storedItems.entrySet());
             sortedItems.sort(Comparator.comparing(e -> e.getKey().getMaterialName()));
-            for (Map.Entry<VirtualInventory.ItemSignature, Long> entry : sortedItems) {
+            for (Map.Entry<ItemSignature, Long> entry : sortedItems) {
                 Material material = entry.getKey().getTemplateRef().getType();
                 long amount = entry.getValue();
                 String formattedAmount = languageManager.formatNumber(amount);
